+++ JPA/src/main/java/com/otp/Xamp/Controller/homePage.java
@@ -1,4 +1,4 @@
-package com.RestAPI.JPA.Controller;
+package com.otp.Xamp.Controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,6 +12,7 @@ public class homePage {
 	public ModelAndView index () {
 	    ModelAndView modelAndView = new ModelAndView();
 	    modelAndView.setViewName("index");
+	    System.out.println("simple message");
 	    return modelAndView;
 	}
