+++ JPA/src/main/java/com/otp/Xamp/Repository/UserRepository.java
@@ -1,8 +1,9 @@
-package com.RestAPI.JPA.Repository;
+package com.otp.Xamp.Repository;
 
-import com.RestAPI.JPA.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.otp.Xamp.Entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
