+++ JPA/src/main/java/com/otp/Xamp/Repository/candidateRepo.java
@@ -1,8 +1,9 @@
-package com.RestAPI.JPA.Repository;
+package com.otp.Xamp.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.RestAPI.JPA.Entity.Candidate;
+
+import com.otp.Xamp.Entity.Candidate;
 
 @Repository
 public interface candidateRepo extends JpaRepository<Candidate, Integer>{
