+++ JPA/src/main/java/com/otp/Xamp/Service/UserService.java
@@ -1,9 +1,10 @@
-package com.RestAPI.JPA.Service;
+package com.otp.Xamp.Service;
 
-import com.RestAPI.JPA.Entity.User;
-import com.RestAPI.JPA.Repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.otp.Xamp.Entity.User;
+import com.otp.Xamp.Repository.UserRepository;
 
 import java.util.List;
 import java.util.Optional;
