+++ JPA/src/main/java/com/otp/Xamp/Service/UserServiceImpl.java
@@ -1,12 +1,13 @@
-package com.RestAPI.JPA.Service;
+package com.otp.Xamp.Service;
 
-import com.RestAPI.JPA.Entity.User;
-import com.RestAPI.JPA.Repository.UserRepository;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.repository.query.FluentQuery;
+
+import com.otp.Xamp.Entity.User;
+import com.otp.Xamp.Repository.UserRepository;
 
 import java.util.List;
 import java.util.Optional;
