+++ JPA/src/main/java/com/otp/Xamp/XampApplication.java
@@ -1,13 +1,13 @@
-package com.RestAPI.JPA;
+package com.otp.Xamp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class JpaApplication {
+public class XampApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(JpaApplication.class, args);
+		SpringApplication.run(XampApplication.class, args);
 	}
 
 }
