+++ JPA/src/test/java/com/RestAPI/JPA/XampApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class JpaApplicationTests {
+class XampApplicationTests {
 
 	@Test
 	void contextLoads() {
